requirements.txt
# Python libraries
pandas
numpy
psycopg2
python-dotenv